// github/types/paths/repos/{owner}/{repo}/labels/{name}.types.ts
export type issuesGetLabel = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesGetLabel_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: label;
          };
        };
        examples: {
          default: unknown;
        };
      };
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Updates a label using the given label name.
 */
export type issuesUpdateLabel = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesUpdateLabel_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The new name of the label. Emoji can be added to label names, using either native emoji or colon-style markup. For example, typing `:strawberry:` will render the emoji ![:strawberry:](https://github.githubassets.com/images/icons/emoji/unicode/1f353.png ":strawberry:"). For a full list of available emoji and codes, see "[Emoji cheat sheet](https://github.com/ikatyang/emoji-cheat-sheet)."
       */
      new_name?: string;
      /**
       * The [hexadecimal color code](http://www.color-hex.com/) for the label, without the leading `#`.
       */
      color?: string;
      /**
       * A short description of the label. Must be 100 characters or fewer.
       */
      description?: string;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: label;
          };
        };
        examples: {
          default: unknown;
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Deletes a label using the given label name.
 */
export type issuesDeleteLabel = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesDeleteLabel_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type issuesGetLabel_Path = { owner: string; repo: string; name: string };

export type issuesUpdateLabel_Path = {
  owner: string;
  repo: string;
  name: string;
};

export type issuesDeleteLabel_Path = {
  owner: string;
  repo: string;
  name: string;
};